releasetools: Store the build.prop file in the OTA zip
This file is often used to read information about the update contained
in the OTA. Place it in the update so it can be used by scripts. Add
build.prop to incrementals too

The file is not added to the updater-script, so it will not be placed
onto the actual system.

Change-Id: I88044796cbe8f199ca02df2840fd944cba2c73fa
Signed-off-by: SagarMakhar <sagarmakhar@gmail.com>